drivers/net/ethernet: Wrapper Cleanup
Some of the Linux function calls got changed in later versions of Linux
to require the caller to be GPLd.  We are in the process of GPLing all
our driver code, but were not yet finished doing so. The wrappers were
a temporary work around until the RTE (the external module we provide
to customers that controls the rest of the hardware) is GPLd.

Signed-off-by: John Jacques <john.jacques@lsi.com>