drivers/net: The work around for Amarillo needs to be done the first time as well!
The original hardware disabled auto negotiation. This was anticipated by the
driver, but the hardware did not work as expected. The PHY status registers
did not return the correctly values. As a work around, until auto negotiation
could be re-enabled, we forced the MAC into half duplex mode. This mitigates
the problem, but does not solve it. Most of the boards have been fixed at
this point.

Signed-off-by: John Jacques <john.jacques@lsi.com>