LSI ACP34XX: Fix interrupts specification for ACP34xx dts files
The MPIC: interrupt-controller #interrupt-cells is <2>, which requires
at least two parameters for a device's "interrupts" list.  Update the
plb device's "interrupts" entries in the acp34xx.dts files to include
an interrupt number and an interrupt level/edge trigger, to fix a
a mismatch with the MPIC: interrupt-controller's #interrupt-cells.

This fixes a failure when the device's interrupt is looked up in
the Open Firmware device tree (e.g. irq_of_parse_and_map).

Signed-off-by: Hong H. Pham <hong.pham@windriver.com>