arch/arm/mach-axxia: Support for DDR Retention Resets
The Axxia hardware makes it possible to reset the system without
losing the contents of memory.  This commit adds the ability to
initiate a reset in such a way that memory contents are not
lost.

Cleanup of the DDR Retention feature for Linux 3.10.

Shutdown devices, and disable hotplug before shuting down
the cores.

Signed-off-by: John Jacques <john.jacques@lsi.com>