i2c-axxia: Fix broken smbus block read
Changed the initial transfer size on block reads from 1 to I2C_SMBUS_BLOCK_MAX.
The size is adjusted when the first byte (block length) is received. Having the
initial size set to 1 could cause the controller to stop the transfter after
the block length byte, if the transfter length register wasn't updated in time.

Signed-off-by: Anders Berg <anders.berg@lsi.com>