arm/axxia: Workaround for cross-cluster sev issue
In revision 1.0 of the AXM55xx device, the events signals across cluster can
get lost if multiple CPUs signals an event roughly at the same point in time.
This may lead to spin-lock deadlock as other clusters may be sitting in a wfe
instruction waiting for a lock to become free.

The workaround is to disable the wfe instruction (effectively turning it into a
NOP) by setting ACTLR[7]=1.

Signed-off-by: Anders Berg <anders.berg@lsi.com>