ARM: axxia: Alloc GIC to be initialized from non-primary core
The condition to only initialize the distributor interface from the primary
core of each cluster could fail when the primary core is not started (e.g. in a
crash kernel started via kexec). In this case, the one and only CPU may be any
core in a cluster.

This change handles this by initializing the GIC distributor interface via the
first CPU calling the init function, by keeping a state which clusters have
been initialized.

Signed-off-by: Anders Berg <anders.berg@lsi.com>