arch/arm/mach-axxia: Added the Redundant Parameters inthe Device Tree
The partitioning of serial flash is what got changed in the device tree.
It now matches the images U-Boot expects (spl binary, parameters (two copies),
U-Boot environment (two copies), and U-Boot images.

This patch fixed the the difference in the Linux device tree for the
serial flash partitioning. It supports fall back images for the U-Boot
parameter file and U-Boot images, It is a new feature.

This change is required to support that feature.

Signed-off-by: John Jacques <john.jacques@lsi.com>