arch/arm/mach-axxia: Move Simulutaion/Emulation WFE/SEV Work-Around
In some simulation and emulation environments, WFE/SEV is either not
supported or doesn't work.  The work-around is to determine, using the
device tree, when running in either of those environments and disable
wfe.  This was originally done when the architecture specific wfe
function first got called, but that can cause lockups because the work-
around reads the device tree.  It is now done in the architecture's
platsmp.c file before enabling the secondary cores.

Signed-off-by: John Jacques <john.jacques@lsi.com>