net: femac: Avoid panic in skb_put()
On very rare occations the femac_rx_packet() will overrun the sk_buff causuing
skb_put() to panic. To current date, we have only had one reported case which
makes it hard to diagnose. This patch is not a proper solution, it is a step on
the way to get some more information on the real issue. To avoid painc from
skb_put(), we make sure there is tailroom in the skb before adding more data to
it, and emit an error log if the buffer was full.

Signed-off-by: Anders Berg <anders.berg@avagotech.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>