arch/arm/mach-axxia: Fixed affinity error on IRQ migration
This patch redesigned the gic_set_affinity function and added
a system work queue to process the affinity requests that are
needed to be processed via the IPI. The reason for the workqueue
is to insure that all interrupts get migrated. When code relies
on and interrupt to be updated and that interrupt is being migrated,
problems arose.

This patch also adds a choice to hotplug. There are two ways to
power down the cpu, either to a low power mode or completely off.
If the power off mode is selected then the a new option is available
to power off the L2 cache as well.

Signed-off-by: John Jacques <john.jacques@lsi.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>